requirements.txt
Django==1.11.2
django-guardian==1.4.8
mysqlclient==1.3.10
pytz==2017.2
six==1.10.0
flake8
